Abstract out file IO to Storage class
